feat: return stable values based on request ip
This obviously relies on stable host ip addresses, but is an okay
start. This requires configuring a dhcp reservation for the host's mac
address in the dhcp server BEFORE the initial boot of the host. It also
requires preserving the file that these mappings are stored in.
Obviously, there's lots of room for improvement here. For example,
adding some coordination with the hypervisor and/or dhcp server to make
this more robust.

authored-by: gpt-5-codex