feat: add v3 endpoints with hostname to ip mapping
This is all fairly hacky and unclean. I could easily blame that on the
author (ai). My goal, for now, is not cleanliness but rather to automate
the last remaining manual step of host and kubernetes cluster
provisioning, which is caused by hosts reverting to dhcp on boot. Not
the end of the world, but it means dns entries point to the wrong place
and machines get "lost" on the default dhcp lease range. I can work
around this by creating dhcp reservations, but that's manual and
therefore annoying to do compared to running a magic script.