feat: add network-config endpoint
Still trying to get a static ip address to be respected by cloud-init.
No matter what I do, the machine ends up inappropriately using dhcp at
some point. It needs to initially, in order to even hit the network data
source where I'm serving dynamic cloud-init content. So, that's fine,
and I can get it to use the static ip after that, but as soon as the
machine reboots it goes right back to dhcp. It's quite clear at this
point that neither I, nor ai, understand cloud-init nor what is going on
during boot.

I also had ai write the first test (shame that it took me this long, and
shame I pawned it off on ai). Moved the single server.js source file
into a src directory in preparation for some sorely needed refactoring
of this giant wall of code.

I have very little hope of any of this (meaning this commit's content)
working. I think I still need to rip out some of the ai code and just
rely on a static file for hostname to ip address mapping. That was the
intent of the vm-ip-config.json file, but ai ran with it down a rabbit
hole. Last time I ran the server, it still displayed some odd
logs/behavior.

Lots of todo items in my mind, but I think I need to wrap up this commit
message...